requirements.txt
uvicorn
fastapi
torch==2.5.1 torchvision torchaudio
coqui-tts
pyinstaller